requirements.txt
black==24.10.0
bcrypt==4.2.1
fastapi==0.115.6
uvicorn==0.34.0
sqlalchemy==2.0.36
python-dotenv==1.0.1
pyjwt==2.10.1
pytest==8.3.4
bcrypt==4.2.1
email_validator==2.2.0
httpx==0.28.1